Refactor : 회원가입 페이지 생성 완료
각각의 버튼 클릭시 data가 보내지는 기능을 추가해야함
