+++ frontend/package.json
@@ -17,6 +17,7 @@
     "react-bootstrap": "^2.7.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
+    "styled-components": "^5.3.6",
     "tailwindcss": "^3.2.4",
     "typescript": "^4.9.4",
     "web-vitals": "^2.1.4"
