+++ frontend/src/App.tsx
@@ -1,11 +1,13 @@
 import React from "react";
-import Login from "./components/login/login";
+import Login from "./routes/login/login";
+import Register from "./routes/register/register";
+import "./index.css";
 import "bootstrap/dist/css/bootstrap.min.css";
 
 function App() {
   return (
     <>
-      <Login />
+      <Register />
     </>
   );
 }
