+++ frontend/src/routes/login/login.tsx
@@ -1,0 +1,11 @@
+import { Button, Container, Row, Col } from "react-bootstrap"; // 꼭 import를 해와야한다
+
+function Login() {
+  return (
+    <Container>
+      
+    </Container>
+  );
+}
+
+export default Login;
