+++ frontend/src/routes/register/register.css
@@ -1,0 +1,6 @@
+.center {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+}
